requirements.txt
Flask==2.3.3
gunicorn
requests
googletrans==3.1.0a0
psycopg2-binary==2.9.9
